refactor: put the jsonapi error to the bottom
